Refactor Google Login Page. Removed Show and move all data to display on List page.
