+++ src/App.js
@@ -5,14 +5,16 @@ import Dashboard from './pages/dashboard';
 import authProvider from './dataProvider/authProvider';
 import dataProvider from './dataProvider/dataProvider';
 import loginPage from './pages/LoginPage'
-import scenes from './pages/scene';
+
 import { CourseList, CourseEdit, CourseCreate } from './pages/courses';
 import { ReferenceExamplesList, ReferenceExamplesEdit, ReferenceExamplesCreate } from './pages/referenceExamples';
 import { UserList, UserEdit, UserCreate } from './pages/users';
 import { SnapshotList, SnapshotShow } from './pages/snapshots';
 import { NotifList, NotifEdit, NotifCreate } from './pages/Notifications';
 import { CollectionList, CollectionShow } from './pages/Collections';
-import { GoogleLoginList, GoogleLoginShow } from './pages/GoogleLogins';
+
+import scenes from './pages/scene';
+import googleLogins from "./pages/googleLogin";
 
 import {
   List                 as CourseIcon,
@@ -21,7 +23,6 @@ import {
   CameraEnhance        as SnapshotIcon,
   Filter               as CollectionIcon,
   PriorityHigh         as NotificationIcon,
-  Face                 as GoogleLoginIcon
 } from "@mui/icons-material";
 
 const App = () => (
@@ -31,7 +32,7 @@ const App = () => (
     <Resource name="courses" list={CourseList} edit={CourseEdit} create={CourseCreate} icon={CourseIcon} />
     <Resource name="referenceExamples" list={ReferenceExamplesList} edit={ReferenceExamplesEdit} create={ReferenceExamplesCreate} icon={RefExIcon} />
     <Resource name="users" list={UserList} edit={UserEdit} create={UserCreate} icon={UserIcon} />
-    <Resource name="googlelogins" list={GoogleLoginList} show={GoogleLoginShow} icon={GoogleLoginIcon} />
+    <Resource name="googlelogins"  {...googleLogins} />
     <Resource name="notifications" list={NotifList} edit={NotifEdit} create={NotifCreate} icon={NotificationIcon} />
     <Resource name="snapshots" list={SnapshotList} show={SnapshotShow} icon={SnapshotIcon} />
   </Admin>
