+++ src/pages/googleLogin/index.js
@@ -1,0 +1,8 @@
+import GoogleLoginIcon from "@mui/icons-material/Face";
+
+import GoogleLoginList from "./GoogleLoginList";
+
+export default {
+    icon: GoogleLoginIcon,
+    list: GoogleLoginList,
+};
